test(config): split config specs from runtime
